Change Estimate to store the values and unit the user actually inputs, instead of a value scaled to milliseconds. This is to make the user interface more intuitive, the previous solution was a bit too clever. The resulting calculated values used in sceduling are cached. Even added some unit tests ;)
svn path=/trunk/koffice/; revision=742899

+++ kplato/libs/kernel/kptcalendar.cpp
@@ -258,7 +258,7 @@ Duration CalendarDay::workDuration() const
     }
     foreach (TimeInterval *i, m_workingIntervals) {
         //kDebug()<<"Interval:"<<i->first<<" -"<<i->second;
-        d += Duration(i->first.msecsTo( i->second ) );
+        d += Duration( (qint64)(i->first.msecsTo( i->second )) );
     }
     return d;
 }
